restore original dev/pass login to match .yml
Seems to have been a mix-up with the original dev/pass login from the parent repo & exponent/exponent login in this repo. The .yml file was not updated so CartoDB expects dev.localhost (and this is the url that has to be visited to access the page, as detailed in the README).